// src/interface/data.interface.tsx
export interface arrayLanesProps {
    title: string;
    items: any;
    color: string;
}

export interface arrayDataProps {
    title: string;
    date: string;
}

export interface InitialStateProps {
    loading: boolean;
    value: string;
    userInfo: {};
    userToken: string;
    error: string;
    success: boolean;
}